p4i integration Integrate 207423 from 4.1 to main: Fixes:    Disappearing text when draw as path onto an image Task:     #100893 RevBy:    Done with Gunnar Details:  Fix short overflows and out of memory handling in grayraster
[git-p4: depot-paths = "//depot/qt/main/": change = 207424]

+++ src/gui/painting/qpaintengine_raster.cpp
@@ -247,7 +247,7 @@ public:
     QDataBuffer<QPointF> m_elements_dev;
     QDataBuffer<QT_FT_Vector> m_points;
     QDataBuffer<char> m_tags;
-    QDataBuffer<short> m_contours;
+    QDataBuffer<int> m_contours;
 
     QPolygonClipper<QRasterFloatPoint, QRasterFloatPoint, qreal> m_clipper;
     QDataBuffer<QPointF> m_polygon_dev;
